+++ lua/obsidian/commands/toggle_checkbox.lua
@@ -4,7 +4,7 @@ local toggle_checkbox = require("obsidian.util").toggle_checkbox
 return function(client)
   local checkboxes = vim.tbl_keys(client.opts.ui.checkboxes)
   table.sort(checkboxes, function(a, b)
-    return client.opts.ui.checkboxes[a].order < client.opts.ui.checkboxes[b].order
+    return (client.opts.ui.checkboxes[a].order or 1000) < (client.opts.ui.checkboxes[b].order or 1000)
   end)
   toggle_checkbox(checkboxes)
 end